Move rewriters into a sub-library
