+++ lib/macroni/Conversion/MacroniRewriters.cpp
@@ -1,4 +1,4 @@
-#include "MacroniRewriters.hpp"
+#include <macroni/Conversion/MacroniRewriters.hpp>
 
 namespace macroni {
     bool is_get_user(macroni::MacroExpansion exp) {
